+++ EssentialFeedTests/FeedAPI/RemoteFeedImageDataLoaderTests.swift
@@ -8,189 +8,138 @@
 import XCTest
 import EssentialFeed
 
-class RemoteFeedImageDataLoader {
-    private let client: HTTPClient
-    init(client: HTTPClient) {
-        self.client = client
-    }
-    public enum Error: Swift.Error {
-        case invalidData
+class RemoteFeedImageDataLoaderTests: XCTestCase {
+    
+    func test_init_doesNotPerformAnyURLRequest() {
+        let (_, client) = makeSUT()
+        XCTAssertTrue(client.requestedURLs.isEmpty)
     }
     
-    private final class HTTPClientTaskWrapper: FeedImageDataLoaderTask {
-        var wrapped: HTTPClientTask?
-        private var completion: ((FeedImageDataLoader.Result) -> Void)?
-        
-        init(_ completion: @escaping (FeedImageDataLoader.Result) -> Void) {
-            self.completion = completion
+    func test_loadImageData_requestDataFromURL() {
+        let (sut, client) = makeSUT()
+        let url = anyURL()
+        sut.loadImageData(from: url) { _ in
         }
-        
-        func complete(with result: FeedImageDataLoader.Result) {
-            completion?(result)
-        }
-        
-        func cancel() {
-            preventFurtherCompletions()
-            wrapped?.cancel()
-        }
-        private func preventFurtherCompletions() {
-            completion = nil
-        }
+        XCTAssertEqual(client.requestedURLs, [url])
     }
     
-    @discardableResult
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-        let task = HTTPClientTaskWrapper(completion)
-        
-        task.wrapped = client.get(from: url) { [weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .success((data, response)):
-                if response.statusCode == 200, !data.isEmpty {
-                    task.complete(with: .success(data))
-                } else {
-                    task.complete(with: .failure(Error.invalidData))
-                }
-            case let .failure(error):
-                task.complete(with: .failure(error))
-            }
+    func test_loadImageDataFromURLTwice_requestsDataFromURLTwice() {
+        let (sut, client) = makeSUT()
+        let url = anyURL()
+        sut.loadImageData(from: url) { _ in
         }
-        return task
+        sut.loadImageData(from: url) { _ in
+        }
+        XCTAssertEqual(client.requestedURLs, [url, url])
     }
-}
     
-    class RemoteFeedImageDataLoaderTests: XCTestCase {
+    func test_loadImageDataFromURL_deliversErrorOnClientError() {
+        let (sut, client) = makeSUT()
+        let clientError = NSError(domain: "a client error", code: 0)
         
-        func test_init_doesNotPerformAnyURLRequest() {
-            let (_, client) = makeSUT()
-            XCTAssertTrue(client.requestedURLs.isEmpty)
-        }
-        
-        func test_loadImageData_requestDataFromURL() {
-            let (sut, client) = makeSUT()
-            let url = anyURL()
-            sut.loadImageData(from: url) { _ in
-            }
-            XCTAssertEqual(client.requestedURLs, [url])
-        }
-        
-        func test_loadImageDataFromURLTwice_requestsDataFromURLTwice() {
-            let (sut, client) = makeSUT()
-            let url = anyURL()
-            sut.loadImageData(from: url) { _ in
-            }
-            sut.loadImageData(from: url) { _ in
-            }
-            XCTAssertEqual(client.requestedURLs, [url, url])
-        }
-        
-        func test_loadImageDataFromURL_deliversErrorOnClientError() {
-            let (sut, client) = makeSUT()
-            let clientError = NSError(domain: "a client error", code: 0)
-            
-            expect(sut, toCompleteWith: .failure(clientError), when: {
-                client.complete(with: clientError)
-            })
-        }
-        
-        func test_loadImageDataFromURL_deliversInvalidDataOnNon200HTTPResponse() {
-            let (sut, client) = makeSUT()
-            let samples = [199, 201, 300, 400, 500]
-            samples.enumerated().forEach { index, errorCode in
-                expect(sut, toCompleteWith: failure(.invalidData), when: {
-                    client.complete(withStatusCode: errorCode, data: anyData(), at: index)
-                })
-            }
-        }
-        
-        func test_loadImageDataFromURL_deliversInvalidDataErrorOn200HTTPResponseWithEmptyData() {
-            let (sut, client) = makeSUT()
+        expect(sut, toCompleteWith: .failure(clientError), when: {
+            client.complete(with: clientError)
+        })
+    }
+    
+    func test_loadImageDataFromURL_deliversInvalidDataOnNon200HTTPResponse() {
+        let (sut, client) = makeSUT()
+        let samples = [199, 201, 300, 400, 500]
+        samples.enumerated().forEach { index, errorCode in
             expect(sut, toCompleteWith: failure(.invalidData), when: {
-                let emptyData = Data()
-                client.complete(withStatusCode: 200, data: emptyData)
+                client.complete(withStatusCode: errorCode, data: anyData(), at: index)
             })
         }
-        
-        func test_loadImageDataFromURL_deliversReceivedNonEmptyDataOn200HTTPResponse() {
-            let (sut, client) = makeSUT()
-            let nonEmptyData = Data("non-empty data".utf8)
-            
-            expect(sut, toCompleteWith: .success(nonEmptyData), when: {
-                client.complete(withStatusCode: 200, data: nonEmptyData)
-            })
-        }
-        
-        func test_loadImageDataFromURL_doesNotDeliverResultAfterSUTInstanceHasBeenDeallocated() {
-            let client = HTTPClientSpy()
-            var sut: RemoteFeedImageDataLoader? = RemoteFeedImageDataLoader(client: client)
-            
-            var capturedeResults = [FeedImageDataLoader.Result]()
-            sut?.loadImageData(from: anyURL(), completion: { result in
-                capturedeResults.append(result)
-            })
-            sut = nil
-            client.complete(withStatusCode: 200, data: anyData())
-            XCTAssertTrue(capturedeResults.isEmpty)
-        }
-        
-        func test_cancelLoadImageDataURLTask_cancelsClientURLRequest() {
-            let (sut, client) = makeSUT()
-            let url = URL(string: "https://a-given-url.com")!
-            let task = sut.loadImageData(from: url) { _ in }
-            XCTAssertTrue(client.cancelledURLs.isEmpty, "Expected no cancelled URL request until task is cancelled")
-            task.cancel()
-            XCTAssertEqual(client.cancelledURLs, [url], "Expected cancelled URL request after task is cancelled")
-        }
-        
-        
-        //    func test_loadImageDataFromURL_doesNotDeliverResultAfterCancellingTask() {
-        //        let (sut, client) = makeSUT()
-        //        let nonEmptyData = Data("non-empty data".utf8)
-        //
-        //        var received = [FeedImageDataLoader.Result]()
-        //        let task = sut.loadImageData(from: anyURL()) {
-        //            received.append($0)
-        //        }
-        //        task.cancel()
-        //
-        //        client.complete(withStatusCode: 404, data: anyData())
-        //        client.complete(withStatusCode: 200, data: nonEmptyData)
-        //        client.complete(with: anyNSError())
-        //
-        //        XCTAssertTrue(received.isEmpty, "Expected no received results after cancelling task")
-        //    }
-        
-        //MARK :- Helpers
-        
-        private func makeSUT(url:URL =  anyURL(), file: StaticString = #filePath, line: UInt = #line) -> (sut: RemoteFeedImageDataLoader, client: HTTPClientSpy) {
-            let client = HTTPClientSpy()
-            let sut = RemoteFeedImageDataLoader(client: client)
-            trackForMemoryLeaks(sut, file: file, line: line)
-            trackForMemoryLeaks(client, file: file, line: line)
-            return (sut, client)
-        }
-        
-        private func expect(_ sut: RemoteFeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-            let url = URL(string: "https://a-given-url.com")!
-            let exp = expectation(description: "Wait for load completion")
-            sut.loadImageData(from: url) { receivedResult in
-                switch (receivedResult, expectedResult) {
-                case let (.success(receivedData), .success(expectedData)):
-                    XCTAssertEqual(receivedData, expectedData, file: file, line: line)
-                    
-                case let (.failure(receivedError as NSError), .failure(expectedError as NSError)):
-                    XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-                default :
-                    XCTFail("Expected result \(expectedResult) got \(receivedResult) instead", file: file, line: line)
-                }
-                exp.fulfill()
-            }
-            action()
-            wait(for: [exp], timeout: 1.0)
-        }
-        
-        private func failure(_ error: RemoteFeedImageDataLoader.Error) -> FeedImageDataLoader.Result {
-            return .failure(error)
-        }
-
     }
+    
+    func test_loadImageDataFromURL_deliversInvalidDataErrorOn200HTTPResponseWithEmptyData() {
+        let (sut, client) = makeSUT()
+        expect(sut, toCompleteWith: failure(.invalidData), when: {
+            let emptyData = Data()
+            client.complete(withStatusCode: 200, data: emptyData)
+        })
+    }
+    
+    func test_loadImageDataFromURL_deliversReceivedNonEmptyDataOn200HTTPResponse() {
+        let (sut, client) = makeSUT()
+        let nonEmptyData = Data("non-empty data".utf8)
+        
+        expect(sut, toCompleteWith: .success(nonEmptyData), when: {
+            client.complete(withStatusCode: 200, data: nonEmptyData)
+        })
+    }
+    
+    func test_loadImageDataFromURL_doesNotDeliverResultAfterSUTInstanceHasBeenDeallocated() {
+        let client = HTTPClientSpy()
+        var sut: RemoteFeedImageDataLoader? = RemoteFeedImageDataLoader(client: client)
+        
+        var capturedeResults = [FeedImageDataLoader.Result]()
+        sut?.loadImageData(from: anyURL(), completion: { result in
+            capturedeResults.append(result)
+        })
+        sut = nil
+        client.complete(withStatusCode: 200, data: anyData())
+        XCTAssertTrue(capturedeResults.isEmpty)
+    }
+    
+    func test_cancelLoadImageDataURLTask_cancelsClientURLRequest() {
+        let (sut, client) = makeSUT()
+        let url = URL(string: "https://a-given-url.com")!
+        let task = sut.loadImageData(from: url) { _ in }
+        XCTAssertTrue(client.cancelledURLs.isEmpty, "Expected no cancelled URL request until task is cancelled")
+        task.cancel()
+        XCTAssertEqual(client.cancelledURLs, [url], "Expected cancelled URL request after task is cancelled")
+    }
+    
+    
+    //    func test_loadImageDataFromURL_doesNotDeliverResultAfterCancellingTask() {
+    //        let (sut, client) = makeSUT()
+    //        let nonEmptyData = Data("non-empty data".utf8)
+    //
+    //        var received = [FeedImageDataLoader.Result]()
+    //        let task = sut.loadImageData(from: anyURL()) {
+    //            received.append($0)
+    //        }
+    //        task.cancel()
+    //
+    //        client.complete(withStatusCode: 404, data: anyData())
+    //        client.complete(withStatusCode: 200, data: nonEmptyData)
+    //        client.complete(with: anyNSError())
+    //
+    //        XCTAssertTrue(received.isEmpty, "Expected no received results after cancelling task")
+    //    }
+    
+    //MARK :- Helpers
+    
+    private func makeSUT(url:URL =  anyURL(), file: StaticString = #filePath, line: UInt = #line) -> (sut: RemoteFeedImageDataLoader, client: HTTPClientSpy) {
+        let client = HTTPClientSpy()
+        let sut = RemoteFeedImageDataLoader(client: client)
+        trackForMemoryLeaks(sut, file: file, line: line)
+        trackForMemoryLeaks(client, file: file, line: line)
+        return (sut, client)
+    }
+    
+    private func expect(_ sut: RemoteFeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
+        let url = URL(string: "https://a-given-url.com")!
+        let exp = expectation(description: "Wait for load completion")
+        sut.loadImageData(from: url) { receivedResult in
+            switch (receivedResult, expectedResult) {
+            case let (.success(receivedData), .success(expectedData)):
+                XCTAssertEqual(receivedData, expectedData, file: file, line: line)
+                
+            case let (.failure(receivedError as NSError), .failure(expectedError as NSError)):
+                XCTAssertEqual(receivedError, expectedError, file: file, line: line)
+            default :
+                XCTFail("Expected result \(expectedResult) got \(receivedResult) instead", file: file, line: line)
+            }
+            exp.fulfill()
+        }
+        action()
+        wait(for: [exp], timeout: 1.0)
+    }
+    
+    private func failure(_ error: RemoteFeedImageDataLoader.Error) -> FeedImageDataLoader.Result {
+        return .failure(error)
+    }
+    
+}
